+++ sync/patches/kamaji-crds/crds/patch.sh
@@ -6,7 +6,7 @@ set -o pipefail
 
 repo_dir=$(git rev-parse --show-toplevel) ; readonly repo_dir
 script_dir=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd ) ; readonly script_dir
-CHART_DIR="${repo_dir}/helm/kamaji" ; readonly CHART_DIR
+CHART_DIR="${repo_dir}/helm/kamaji/charts/kamaji-crds" ; readonly CHART_DIR
 
 cd "${repo_dir}"
 
@@ -14,8 +14,8 @@ readonly script_dir_rel=".${script_dir#"${repo_dir}"}"
 
 set -x
 
-git apply "${script_dir_rel}/_values.patch"
-
-cp -av "${script_dir_rel}/manifests/values.schema.json" "${CHART_DIR}/values.schema.json"
+git apply "${script_dir_rel}/_kamaji.clastix.io_datastores.yaml"
+git apply "${script_dir_rel}/_kamaji.clastix.io_kubeconfiggenerators.yaml"
+git apply "${script_dir_rel}/_kamaji.clastix.io_tenantcontrolplanes.yaml"
 
 { set +x; } 2>/dev/null