+++ sync/sync.sh
@@ -24,11 +24,11 @@ mv helm/kamaji-crds "$kamaji_chart_dir/charts/"
  ./sync/patches/kamaji/chart/patch.sh
  ./sync/patches/kamaji/values/patch.sh
  ./sync/patches/kamaji/helpers/patch.sh
+ ./sync/patches/kamaji/kube-linter/patch.sh
 
 # Patches for Kamaji CRDs
  ./sync/patches/kamaji-crds/chart/patch.sh
+ ./sync/patches/kamaji-crds/crds/patch.sh
  ./sync/patches/kamaji-crds/crd-conversion/patch.sh
  ./sync/patches/kamaji-crds/helpers/patch.sh
  ./sync/patches/kamaji-crds/values/patch.sh
-
-helm dependency update ./helm/kamaji