+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add Helm `keep` annotation to all CRDs.
+- Added values schema to kamaji chart.
+
 ### Changed
 
 - Simplify chart structure